+++ Mods ChoGGi/View Colony Map/Code/Script.lua
@@ -23,7 +23,7 @@ local image_str = string.format("%sMaps/%s.png",Mods.ChoGGi_MapImagesPack.env.Cu
 -- tell people how to get my library mod (if needs be)
 function OnMsg.ModsReloaded()
 	-- version to version check with
-	local min_version = 53
+	local min_version = 55
 	local idx = table.find(ModsLoaded,"id","ChoGGi_Library")
 	local p = Platform